pyproject.toml
[project]
name = "mcpserver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "defusedxml>=0.7.1",
    "lxml>=6.0.2",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "openpyxl>=3.1.5",
    "palettable>=3.3.3",
    "pandas>=2.3.3",
    "piexif>=1.1.3",
    "pillow>=12.1.1",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "scienceplots>=2.2.1",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
]
